refactor: split SetupCommand and RunCommand into focused modules
Both files had grown past the 500-line max-lines lint limit and carried
eslint-disable suppressions. Split them along their natural seams so each
module fits the limit and the suppressions are gone:

- commands/setup/SetupTypes.ts: shared setup types, choice catalogs, and
  the settings-issue formatter
- commands/setup/SetupStarterSource.ts: starter layout resolution, git
  setup-source sync, starter/user default-profile settings handling
- commands/setup/SetupSourceImport.ts: copy/symlink import application and
  starter file copying
- commands/setup/SetupPrompts.ts: interactive prompts, readline fallbacks,
  and setup profile discovery
- commands/setup/SetupSourceLaunch.ts: post-import launch actions and
  messaging
- commands/run/RunProfileResolution.ts: resolved-profile loading, profile
  source materialization, and launch layer assembly
- commands/run/RunLaunchSummary.ts: launch summary rendering

SetupCommand.ts and RunCommand.ts keep orchestration plus their public
exports (re-exported from the new modules), so no importer changes. All
code moved verbatim including v8 ignore ranges; 274/274 tests pass with
coverage thresholds intact and no test modifications.
